Revert "Fix Workflows"
