+++ .github/workflows/lint-ts.yml
@@ -1,4 +1,4 @@
-name: Lint Code (ESLint + Prettier)
+name: Lint TS (eslint, prettier)
 
 on:
   push:
@@ -8,26 +8,27 @@ on:
 
 jobs:
   lint:
-    name: Lint Code
+    name: Lint TS (eslint, prettier)
     runs-on: ubuntu-latest
 
     steps:
-      - name: 📦 Checkout repository
+      - name: 📦 Checkout project repo
         uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
 
       - name: 📦 Setup Node + npm + install deps
         uses: ./.github/actions/setup-node-npm-install
 
-      - name: 🎯 Run ESLint
+      - name: 🏃‍♂️ Run ESLint PR
+        if: github.event_name == 'pull_request'
+        uses: reviewdog/action-eslint@v1
+        with:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          reporter: github-pr-review
+          eslint_flags: "."
+          tool_name: eslint
+
+      - name: 🏃‍♂️ Run ESLint Commit
+        if: github.event_name != 'pull_request'
         run: npm run lint
-
-      - name: 🎨 Run Prettier Check
-        run: npm run prettier:check || true
-
-      - name: ✅ Output Success Message
-        if: success()
-        run: echo "✅ Linting completed successfully!"
-
-      - name: ❌ Output Failure Message
-        if: failure()
-        run: echo "❌ Linting failed. Check logs for details."
