+++ android/app/src/main/AndroidManifest.xml
@@ -4,10 +4,6 @@
   <uses-permission android:name="android.permission.SYSTEM_ALERT_WINDOW"/>
   <uses-permission android:name="android.permission.VIBRATE"/>
   <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
-  <uses-permission android:name="android.permission.RECORD_AUDIO"/>
-
-  <uses-feature android:name="android.hardware.microphone" android:required="false"/>
-
   <queries>
     <intent>
       <action android:name="android.intent.action.VIEW"/>
