+++ android/build.gradle
@@ -36,18 +36,6 @@ allprojects {
 
         google()
         mavenCentral()
-
-        maven {
-            url 'https://api.mapbox.com/downloads/v2/releases/maven'
-            authentication {
-                basic(BasicAuthentication)
-            }
-            credentials {
-                username = "mapbox"
-                password = project.hasProperty('MAPBOX_DOWNLOADS_TOKEN') ? MAPBOX_DOWNLOADS_TOKEN : ""
-            }
-        }
-
         maven { url 'https://www.jitpack.io' }
     }
 }
