+++ android/gradle.properties
@@ -22,8 +22,6 @@ org.gradle.jvmargs=-Xmx2048m -XX:MaxMetaspaceSize=512m
 # https://developer.android.com/topic/libraries/support-library/androidx-rn
 android.useAndroidX=true
 
-android.enableJetifier=true
-
 # Enable AAPT2 PNG crunching
 android.enablePngCrunchInReleaseBuilds=true
 
@@ -56,5 +54,3 @@ EX_DEV_CLIENT_NETWORK_INSPECTOR=true
 
 # Use legacy packaging to compress native libraries in the resulting APK.
 expo.useLegacyPackaging=false
-
-MAPBOX_DOWNLOADS_TOKEN=sk.eyJ1Ijoicnd6IiwiYSI6ImNtNm9peDZhdzE4NmQya3E0azV4dmYxenMifQ.5SH51Urj6KLeo-SHYbRTPw
