--- constants/concordiaBuildings.js
@@ -1,890 +1,0 @@
-import {theme} from "@/constants/theme";
-
-export const concordiaBuildingsGeoJSON = {
-    type: "FeatureCollection",
-    features: [
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.57833901279078, 45.49737287972932],
-                        [-73.57903396234046, 45.49770870560358],
-                        [-73.57954373346617, 45.4971647489216],
-                        [-73.57884914044712, 45.49683023814251],
-                        [-73.57833901279078, 45.49737287972932]
-                    ]
-                ]
-            },
-            properties: {
-                name: "H - Henry F. Hall Building",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.5788, 45.4973]
-            }
-        },
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.57729457081123, 45.49689254357292],
-                        [-73.5775555284645, 45.49661429373785],
-                        [-73.5776007245035, 45.496635916100274],
-                        [-73.577649714045, 45.49658239201751],
-                        [-73.57746572181343, 45.49649623135727],
-                        [-73.57769932901236, 45.49625603041722],
-                        [-73.5785668449951, 45.49666799276964],
-                        [-73.57853130572823, 45.496705594685636],
-                        [-73.57857910863052, 45.49672864894097],
-                        [-73.57841436054142, 45.496889961136795],
-                        [-73.5783721157506, 45.49686881013211],
-                        [-73.57833039439095, 45.496912746791004],
-                        [-73.57829217291354, 45.49689535597255],
-                        [-73.57824724591566, 45.49694000806409],
-                        [-73.57829509600118, 45.496963909731356],
-                        [-73.57825687452204, 45.49700057140742],
-                        [-73.57829509599945, 45.49701796219337],
-                        [-73.57805772050071, 45.49725908253964],
-                        [-73.5778169922482, 45.49714063758756],
-                        [-73.57784783765103, 45.497107266147346],
-                        [-73.57779955788186, 45.49708517518128],
-                        [-73.57780626340421, 45.497076714809694],
-                        [-73.57779218180728, 45.49707013451979],
-                        [-73.57774745425716, 45.49711514251558],
-                        [-73.57758664205892, 45.49703908185561],
-                        [-73.57763057384606, 45.496993073045694],
-                        [-73.57762051556332, 45.49698790281031],
-                        [-73.57761179838494, 45.49699730323793],
-                        [-73.577568212493, 45.49697709231659],
-                        [-73.57756754194082, 45.49697662229509],
-                        [-73.57753870819694, 45.49700764370647],
-                        [-73.57729457081123, 45.49689254357292]
-                    ]
-                ]
-            },
-            properties: {
-                name: "Webster Library, Concordia University",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.5778, 45.4970]
-            }
-        },
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.63922806624221, 45.459584199340476],
-                        [-73.6393621766782, 45.45963640633795],
-                        [-73.63932931962138, 45.45967873630042],
-                        [-73.6399210297213, 45.459908052183806],
-                        [-73.63994181683886, 45.45992075111469],
-                        [-73.63994584015194, 45.45992028078398],
-                        [-73.63997266223913, 45.459927806075065],
-                        [-73.63995992174772, 45.45994191599315],
-                        [-73.63996662726952, 45.45994567863738],
-                        [-73.63997132113478, 45.45993250938149],
-                        [-73.64006922175305, 45.459964962185126],
-                        [-73.6400947027359, 45.45993674235694],
-                        [-73.64019729721943, 45.459993181999195],
-                        [-73.64020333218903, 45.45998894902798],
-                        [-73.64023954200675, 45.45987089603626],
-                        [-73.64022881317187, 45.45986619272469],
-                        [-73.64034101830767, 45.459714570778644],
-                        [-73.64015795756255, 45.459647783532496],
-                        [-73.64017338026268, 45.45962473721091],
-                        [-73.63967908522984, 45.45943169128342],
-                        [-73.63965293369483, 45.45945520801854],
-                        [-73.63954832755475, 45.45941052621344],
-                        [-73.63959928952043, 45.459348441962284],
-                        [-73.63945310914521, 45.45929294234659],
-                        [-73.63922806624221, 45.459584199340476]
-                    ]
-                ]
-            },
-            properties: {
-                name: "Physical Services Building (PS)",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.6397, 45.4598]
-            }
-        },
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.64086832135675, 45.45950040615849],
-                        [-73.641070310106, 45.45925163837166],
-                        [-73.64114332951301, 45.45927957491578],
-                        [-73.64114996764093, 45.45927026273594],
-                        [-73.64146385626069, 45.459393981571175],
-                        [-73.64145532152479, 45.45940462403893],
-                        [-73.64154825531554, 45.459441872660314],
-                        [-73.64134721487025, 45.45969063960786],
-                        [-73.64086832135675, 45.45950040615849]
-                    ]
-                ]
-            },
-            properties: {
-                name: "Hingston Hall A (HA)",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.6411, 45.4595]
-            }
-        },
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.6418315606613, 45.45982994727652],
-                        [-73.64191135637073, 45.459727885253365],
-                        [-73.64188051097044, 45.459716126943896],
-                        [-73.64201663306537, 45.45953928167301],
-                        [-73.64224596191092, 45.45962723404424],
-                        [-73.64210715760966, 45.45980313837515],
-                        [-73.64207832386592, 45.45979279107676],
-                        [-73.641998467557, 45.45989466469426],
-                        [-73.6418315606613, 45.45982994727652]
-                    ]
-                ]
-            },
-            properties: {
-                name: "Hingston Hall C (HC)",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.6419, 45.4598]
-            }
-        }, {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                "coordinates": [
-                    [
-                        [-73.64197954429811, 45.459556514714556],
-                        [-73.64200304153985, 45.459526811093816],
-                        [-73.64194202129148, 45.45950235372016],
-                        [-73.64205601516207, 45.459357960932486],
-                        [-73.64202718141833, 45.459289291920044],
-                        [-73.64180388754446, 45.459202279691944],
-                        [-73.6419265985934, 45.45904612768666],
-                        [-73.64215123357368, 45.459133610489864],
-                        [-73.64202718141833, 45.459289291920044],
-                        [-73.64202718141833, 45.459289291920044],
-                        [-73.64207009675991, 45.4593631346187],
-                        [-73.64208954277312, 45.45934008818076],
-                        [-73.64214519860406, 45.45936125327718],
-                        [-73.64230210781416, 45.45916324173116],
-                        [-73.64224444032669, 45.45914019521151],
-                        [-73.64227126241389, 45.459105390245604],
-                        [-73.64190648202015, 45.45896240745426],
-                        [-73.64187898938077, 45.45899674216999],
-                        [-73.64180992250623, 45.458970403211865],
-                        [-73.64149476296052, 45.459368308310346]
-                    ]
-                ]
-            },
-            "properties": {
-                name: "Hingston Hall B (HB)",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.6419, 45.4595]
-            }
-        },
-
-
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.63784440790752, 45.45910190948336],
-                        [-73.63791414533424, 45.45912824838009],
-                        [-73.63791950975167, 45.459121663657065],
-                        [-73.63794364963015, 45.4591306000667],
-                        [-73.6379389557649, 45.45913812546319],
-                        [-73.63813140424054, 45.459212909036204],
-                        [-73.63819644780199, 45.4591329517532],
-                        [-73.63866361649684, 45.45931269267443],
-                        [-73.63867032201865, 45.45930234528602],
-                        [-73.63913524485623, 45.459483389385454],
-                        [-73.63923113381797, 45.45936157265225],
-                        [-73.63920833504386, 45.4593516956083],
-                        [-73.63922107553527, 45.45933570420014],
-                        [-73.63917883074794, 45.45931830178021],
-                        [-73.63922107553527, 45.459264213143435],
-                        [-73.63932434059521, 45.4593051323758],
-                        [-73.63938066697833, 45.4592341116168],
-                        [-73.6393317166692, 45.45921482781914],
-                        [-73.6394182179004, 45.45910759097022],
-                        [-73.6391245160242, 45.45899094714514],
-                        [-73.63904002644952, 45.459095362190325],
-                        [-73.63899711111002, 45.459077959696224],
-                        [-73.63895151356178, 45.45913581120992],
-                        [-73.63898504117077, 45.45914945098906],
-                        [-73.63892737366666, 45.45922000152122],
-                        [-73.63877985218708, 45.45916309076799],
-                        [-73.63878521660452, 45.45915650604904],
-                        [-73.63888714053586, 45.45919601435125],
-                        [-73.63896961845401, 45.45909065881716],
-                        [-73.63886501231393, 45.459049739427186],
-                        [-73.63902021359857, 45.45885426725562],
-                        [-73.63897528660252, 45.45883639434734],
-                        [-73.63898132157213, 45.45882980959021],
-                        [-73.63846767854874, 45.458629914809734],
-                        [-73.63839928222639, 45.45871739825871],
-                        [-73.63836374296085, 45.45870375837506],
-                        [-73.63823700859575, 45.45886602574162],
-                        [-73.63820213988238, 45.458851915554305],
-                        [-73.6381753177907, 45.45888719101661],
-                        [-73.63812569692938, 45.45886931811872],
-                        [-73.63810021593997, 45.458901771534634],
-                        [-73.63804791286994, 45.458882487623285],
-                        [-73.63802615147138, 45.45891098954367],
-                        [-73.63800067048854, 45.45890158276001],
-                        [-73.63789472321865, 45.4590375106347],
-                        [-73.63791483978405, 45.45904644705766],
-                        [-73.63788868824903, 45.45908125205995],
-                        [-73.63786790113146, 45.45907419699366],
-                        [-73.63784440790752, 45.45910190948336]
-                    ]
-                ]
-            },
-            properties: {
-                name: "Vanier Library",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.6385, 45.4592]
-            }
-        },
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.58013564193502, 45.496676493770636],
-                        [-73.58006389285175, 45.49665252254054],
-                        [-73.58005450512123, 45.49666474316893],
-                        [-73.5796675965134, 45.49652984608557],
-                        [-73.57979030756954, 45.496356876525695],
-                        [-73.57980371861314, 45.496360636739084],
-                        [-73.57989223150089, 45.496238899703584],
-                        [-73.57993313518388, 45.49625300053198],
-                        [-73.5799961670888, 45.49616416525411],
-                        [-73.5800910877141, 45.49619848694641],
-                        [-73.58010718096641, 45.49617498554103],
-                        [-73.58044245705639, 45.49629249246985],
-                        [-73.58034992085554, 45.496420809756025],
-                        [-73.58032242821422, 45.49641187926025],
-                        [-73.58013564193502, 45.496676493770636]
-                    ]
-                ]
-            },
-            properties: {
-                name: "Concordia University ER Building",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.5799, 45.4965]
-            }
-        },
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.57957028959329, 45.49653585302086],
-                        [-73.57923297863918, 45.49638213955237],
-                        [-73.57944828350351, 45.496163027264664],
-                        [-73.57967475232375, 45.49627202455517],
-                        [-73.5795838458255, 45.49636704766195],
-                        [-73.57968272657797, 45.49641903082307],
-                        [-73.57957028959329, 45.49653585302086]
-                    ]
-                ]
-            },
-            properties: {
-                name: "Concordia University Learning Square (LS)",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.5794, 45.4964]
-            }
-        },
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.6419058231602, 45.45872621438656],
-                        [-73.64211905875344, 45.458444009269286],
-                        [-73.64171806854984, 45.45829255860727],
-                        [-73.6415048329566, 45.45857664585032],
-                        [-73.6419058231602, 45.45872621438656]
-                    ]
-                ]
-            },
-            properties: {
-                name: "Concordia University - HU Building",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.6417, 45.4585]
-            }
-        },
-
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.5789365603044, 45.494260392494766],
-                        [-73.57910621000592, 45.494364272164106],
-                        [-73.57926915418564, 45.49424723125601],
-                        [-73.5790901167536, 45.49413677073811],
-                        [-73.5789365603044, 45.494260392494766]
-                    ]
-                ]
-            },
-            properties: {
-                name: "CL Building",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.5790, 45.4942]
-            }
-        },
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.57919934449112, 45.495517731161506],
-                        [-73.57896226849431, 45.49543929320034],
-                        [-73.57852621287267, 45.4951870152177],
-                        [-73.57873609571011, 45.495004640777246],
-                        [-73.57879175154105, 45.495038013463514],
-                        [-73.57882460859787, 45.4950065209291],
-                        [-73.57917195462709, 45.495166803644864],
-                        [-73.57911562827067, 45.4952222679425],
-                        [-73.57936641478597, 45.4953585782633],
-                        [-73.57919934449112, 45.495517731161506]
-                    ]
-                ]
-            },
-            properties: {
-                name: "John Molson International Committee",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.5789, 45.4953]
-            }
-        },
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.57849834929043, 45.49586709695727],
-                        [-73.57876388795368, 45.49559541851367],
-                        [-73.57801957500567, 45.49524571283998],
-                        [-73.57800079954463, 45.495264514275306],
-                        [-73.57791496886558, 45.49524571283998],
-                        [-73.57791362776122, 45.495235372047865],
-                        [-73.57774263695535, 45.495353821007505],
-                        [-73.57769718592851, 45.49535433097103],
-                        [-73.57760886228253, 45.49544630898204],
-                        [-73.57849834929043, 45.49586709695727]
-                    ]
-                ]
-            },
-            properties: {
-                name: "EV - Engineering & Visual Arts",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.5783, 45.4956]
-            }
-        },
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.5777848416668, 45.494912049874],
-                        [-73.57771242203137, 45.49486974637085],
-                        [-73.5777043754052, 45.494876796956945],
-                        [-73.57763396742631, 45.49483684362425],
-                        [-73.57762525024796, 45.4948415440178],
-                        [-73.57755014840382, 45.494798770422],
-                        [-73.57754009012112, 45.4948062910566],
-                        [-73.57746498827696, 45.49476398747403],
-                        [-73.57745291833771, 45.4947743283527],
-                        [-73.5773094201712, 45.49469207131071],
-                        [-73.57730003244069, 45.49470100208109],
-                        [-73.57721755452256, 45.49465493809242],
-                        [-73.57751930301247, 45.494397355089625],
-                        [-73.57803898098017, 45.494695361593784],
-                        [-73.5777848416668, 45.494912049874]
-                    ]
-                ]
-            },
-            properties: {
-                name: "Faubourg Tower",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.5776, 45.4948]
-            }
-        },
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.64252381250978, 45.45816553144933],
-                        [-73.6423997603565, 45.458117556202],
-                        [-73.64241518305663, 45.45809921271428],
-                        [-73.64226833212923, 45.4580427711762],
-                        [-73.64223949838548, 45.45805500018093],
-                        [-73.64208057751884, 45.457994795824426],
-                        [-73.64207856583967, 45.457969397091695],
-                        [-73.64200279344334, 45.45794023557027],
-                        [-73.64205308486432, 45.457877209002085],
-                        [-73.6419967584812, 45.45785651369652],
-                        [-73.64202358056839, 45.457822178286456],
-                        [-73.64204839099905, 45.45783205559836],
-                        [-73.64210203517345, 45.457764325424975],
-                        [-73.64213422167809, 45.45777608414148],
-                        [-73.64217914867413, 45.45771964227995],
-                        [-73.64220261800044, 45.457728578911784],
-                        [-73.64224218057906, 45.45767731084673],
-                        [-73.6422710143228, 45.45768765853335],
-                        [-73.64230320083661, 45.45764720847269],
-                        [-73.64229113089738, 45.45764203462589],
-                        [-73.64229917752354, 45.457633568330166],
-                        [-73.64231795298458, 45.4576406235767],
-                        [-73.64233337568471, 45.45762227993384],
-                        [-73.64235282169793, 45.45762980553161],
-                        [-73.64238567875475, 45.45758841473133],
-                        [-73.64238165544168, 45.45758653333059],
-                        [-73.64239506648528, 45.45757007107149],
-                        [-73.64257276281296, 45.45763733112843],
-                        [-73.642558010665, 45.45765520441678],
-                        [-73.64256538673897, 45.45765755616481],
-                        [-73.64250906035586, 45.45772998995657],
-                        [-73.64253454133869, 45.45773892658678],
-                        [-73.64250034317753, 45.457780787625175],
-                        [-73.64253320023434, 45.457793487035154],
-                        [-73.64249229655137, 45.457844754994596],
-                        [-73.64252113029511, 45.457856043346375],
-                        [-73.64247620330646, 45.45791248507531],
-                        [-73.64251509533288, 45.45792659549777],
-                        [-73.64251241312417, 45.45793176931842],
-                        [-73.64253655300264, 45.457941176263866],
-                        [-73.64255599901587, 45.45791577750752],
-                        [-73.64268139227352, 45.45796469362064],
-                        [-73.64252381250978, 45.45816553144933]
-                    ]
-                ]
-            },
-            properties: {
-                name: "St Ignatius Church (SI)",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.6424, 45.4579]
-            }
-        },
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.64056935987028, 45.45717383115564],
-                        [-73.64045201323879, 45.45713008825069],
-                        [-73.64044195495609, 45.45714231745337],
-                        [-73.64016568745795, 45.45704072092006],
-                        [-73.64034740710242, 45.4567989580616],
-                        [-73.64061093410915, 45.45689585147113],
-                        [-73.64062970957018, 45.45687139295627],
-                        [-73.64068871816202, 45.45689349969132],
-                        [-73.6406692721488, 45.45691842855237],
-                        [-73.64074035067988, 45.45694570918012],
-                        [-73.64056935987028, 45.45717383115564]
-                    ]
-                ]
-            },
-            properties: {
-                name: "Centre for Structural and Functional Genomics (GE)",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.6405, 45.4571]
-            }
-        },
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.64141291681526, 45.4583266742855],
-                        [-73.64128299916901, 45.458277451914206],
-                        [-73.64133800080026, 45.458208274454904],
-                        [-73.6412611881773, 45.45817900704274],
-                        [-73.64120144502611, 45.45825550138342],
-                        [-73.64103833674032, 45.45819364075073],
-                        [-73.64092074704591, 45.45833931245497],
-                        [-73.6408619521987, 45.45831669678127],
-                        [-73.64111366268845, 45.45799788500315],
-                        [-73.64106404182712, 45.457978600782646],
-                        [-73.64113243814948, 45.45789252723035],
-                        [-73.64116931851939, 45.457907108005294],
-                        [-73.64147132204852, 45.457525041839446],
-                        [-73.64076414331024, 45.45725002585494],
-                        [-73.6407688371755, 45.45724626303085],
-                        [-73.6406548433049, 45.45720204982869],
-                        [-73.64082918921345, 45.45698358294124],
-                        [-73.64093580701007, 45.45702450383063],
-                        [-73.64095860578418, 45.45699675288589],
-                        [-73.64101292051076, 45.457017448506996],
-                        [-73.64099347449753, 45.45704143660375],
-                        [-73.64129589357474, 45.45715902516967],
-                        [-73.64130394020091, 45.45714867738605],
-                        [-73.64138239480596, 45.45717878002402],
-                        [-73.64139178253649, 45.45716843224404],
-                        [-73.64143335677164, 45.45718442426686],
-                        [-73.64141324020623, 45.45720888264596],
-                        [-73.6420042652158, 45.45743837664774],
-                        [-73.64184534434325, 45.457641568102055],
-                        [-73.64192514005268, 45.457673081525336],
-                        [-73.64141291681526, 45.4583266742855]
-                    ]
-                ]
-            },
-            properties: {
-                name: "SP Building, Loyola Campus, Concordia University",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.6415, 45.4579]
-            }
-        },
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.64004439162848, 45.45762096373572],
-                        [-73.64015033888317, 45.457483621396335],
-                        [-73.64002695727011, 45.4574347048659],
-                        [-73.63994380879177, 45.45744505259704],
-                        [-73.63995453762767, 45.45746198524387],
-                        [-73.63981908607421, 45.45747891788564],
-                        [-73.63977214741708, 45.45742811994511],
-                        [-73.63976343023789, 45.45739566568142],
-                        [-73.63976244833313, 45.45738451989099],
-                        [-73.63976320679176, 45.457358451253086],
-                        [-73.63976927446085, 45.45733185058889],
-                        [-73.63980188818208, 45.45727971325232],
-                        [-73.63982767577559, 45.457259496721065],
-                        [-73.6398853186317, 45.45722917191064],
-                        [-73.6398853186317, 45.457229703925],
-                        [-73.63991869081279, 45.45722491579516],
-                        [-73.63999150284155, 45.45721108341904],
-                        [-73.64001729043508, 45.45721374349162],
-                        [-73.64007265791527, 45.457305248103594],
-                        [-73.64004888672184, 45.45731045875308],
-                        [-73.64004888672247, 45.45731097981874],
-                        [-73.64007562931519, 45.45736048096452],
-                        [-73.64020711373074, 45.45741050313099],
-                        [-73.64039399613314, 45.457176735690744],
-                        [-73.64065780881876, 45.457278810421684],
-                        [-73.64063836280366, 45.45730279840728],
-                        [-73.6407168174163, 45.45733384166757],
-                        [-73.6405008995917, 45.457596297645956],
-                        [-73.64063165727944, 45.45765038788268],
-                        [-73.6404848063367, 45.45783053151701],
-                        [-73.64029101673793, 45.45775386471541],
-                        [-73.64031515661875, 45.457725173431534],
-                        [-73.64004439162848, 45.45762096373572]
-                    ]
-                ]
-            },
-            properties: {
-                name: "Communications & Journalism Building",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.6403, 45.4575]
-            }
-        },
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                "coordinates": [
-                    [
-                        [-73.63973009336604, 45.45829583364624],
-                        [-73.63940889887184, 45.45817354417912],
-                        [-73.63947394243328, 45.45809123381152],
-                        [-73.63954032709911, 45.458118043487316],
-                        [-73.63954837372528, 45.45810863658498],
-                        [-73.63958257188644, 45.4581199248676],
-                        [-73.6396623676023, 45.45801409712429],
-                        [-73.63965432106513, 45.45800516055064],
-                        [-73.63959732413083, 45.45798305428342],
-                        [-73.639637557414, 45.45791720575655],
-                        [-73.63970662428853, 45.457938371356435],
-                        [-73.63971936477997, 45.45794025274272],
-                        [-73.63980117214591, 45.45783442466731],
-                        [-73.63977233840217, 45.45782125493737],
-                        [-73.63978105558051, 45.45781137763789],
-                        [-73.63973746968882, 45.45779444512049],
-                        [-73.63979781918583, 45.45771730803014],
-                        [-73.64009286214502, 45.457830191537924],
-                        [-73.64003251244983, 45.45790497673743],
-                        [-73.63997484496656, 45.45788287043296],
-                        [-73.63987962656144, 45.4580042198417],
-                        [-73.63988901429196, 45.45800845295577],
-                        [-73.63987426214399, 45.458026326100635],
-                        [-73.63991047195189, 45.45803996612843],
-                        [-73.63993192962168, 45.45801127503167],
-                        [-73.63996210446977, 45.458021622642016],
-                        [-73.63998020937765, 45.4579962239587],
-                        [-73.64022227875214, 45.45809029270129],
-                        [-73.64022563151303, 45.45808370788659],
-                        [-73.6405307327957, 45.45819894238607],
-                        [-73.64052469782607, 45.45820599755167],
-                        [-73.6407620732978, 45.45829724428051],
-                        [-73.64071580519737, 45.45835603719767],
-                        [-73.64077749599793, 45.45837861366159],
-                        [-73.6408183996809, 45.45833157935158],
-                        [-73.64097306450022, 45.45839130452468],
-                        [-73.64104280192694, 45.458300058947056],
-                        [-73.6411675246324, 45.458346152597],
-                        [-73.64110048107094, 45.458428992250984],
-                        [-73.64119301727178, 45.45846426792686],
-                        [-73.64121112218064, 45.45844310252399],
-                        [-73.64124129702873, 45.45845627210892],
-                        [-73.64128018905517, 45.45840359375071],
-                        [-73.64134120934034, 45.458425229512504],
-                        [-73.64123460183099, 45.45856021765742],
-                        [-73.6412470128746, 45.45856586175391],
-                        [-73.64112597237784, 45.45872765894625],
-                        [-73.64110920857334, 45.458722485206016],
-                        [-73.64109512697756, 45.458742239484374],
-                        [-73.64099856746366, 45.45870461228152],
-                        [-73.64100795519417, 45.458685798670646],
-                        [-73.64076253304551, 45.45859925597876],
-                        [-73.64084165820275, 45.45850848007883],
-                        [-73.64071693549727, 45.45845815348068],
-                        [-73.64073705205758, 45.45843322533396],
-                        [-73.64067469070483, 45.45840876752091],
-                        [-73.640649209722, 45.45843981012797],
-                        [-73.64041217532073, 45.45835091534246],
-                        [-73.64040882255983, 45.45835750014608],
-                        [-73.6401043918345, 45.458240854937204],
-                        [-73.64010707404322, 45.45823427011997],
-                        [-73.63986433412586, 45.45813784949164],
-                        [-73.6399099316741, 45.45808564118128],
-                        [-73.63984354700827, 45.45806165356301],
-                        [-73.63983348872559, 45.45807435289161],
-                        [-73.63982410100074, 45.45807153081883],
-                        [-73.63972888258834, 45.45819617223298],
-                        [-73.63978520897146, 45.45821733772851]
-                    ]
-                ]
-            },
-            properties: {
-                name: "Loyola Jesuit Hall and Conference Centre (RF), Central Building (CC), Administration Building (AD)",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.6398, 45.4582]
-            }
-        },
-
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.63957049575275, 45.45865344331325],
-                        [-73.63949338225204, 45.45862212986694],
-                        [-73.63948533562588, 45.458633888383705],
-                        [-73.6394182804079, 45.458608489984446],
-                        [-73.63941291599045, 45.45861789680033],
-                        [-73.63937201230748, 45.45860237555329],
-                        [-73.63935927181605, 45.45861883748182],
-                        [-73.6393545779508, 45.45861836714109],
-                        [-73.6393337908332, 45.458643765535885],
-                        [-73.63929489880678, 45.45862871463663],
-                        [-73.63931501537219, 45.45860237555329],
-                        [-73.63930361598511, 45.458597201803325],
-                        [-73.63931568592437, 45.458579799186335],
-                        [-73.6392875227328, 45.45857039236412],
-                        [-73.6392942282546, 45.45856239656397],
-                        [-73.63913463683576, 45.458501252172556],
-                        [-73.6391272607414, 45.458510659005124],
-                        [-73.63909574478895, 45.45851677344622],
-                        [-73.63906624049304, 45.45850501490505],
-                        [-73.63905685276251, 45.458481027473436],
-                        [-73.63906422883649, 45.4584753833704],
-                        [-73.63904813558418, 45.45846879858282],
-                        [-73.63905618221033, 45.45845845105789],
-                        [-73.63900119693157, 45.45843681531787],
-                        [-73.63901527852735, 45.45841941265136],
-                        [-73.63898309202273, 45.4584076540899],
-                        [-73.63896565766605, 45.45842928984113],
-                        [-73.6389334711614, 45.4584175312817],
-                        [-73.63903874786524, 45.45828019112449],
-                        [-73.6390709343699, 45.45829242005601],
-                        [-73.63905551166975, 45.45831499653802],
-                        [-73.63908702762221, 45.45832675511882],
-                        [-73.63909909756144, 45.45830888207505],
-                        [-73.63915542394456, 45.45833051786417],
-                        [-73.63916212946636, 45.458323933059656],
-                        [-73.63918023437522, 45.45833051786417],
-                        [-73.63916212946636, 45.45832440340288],
-                        [-73.63918023437522, 45.45833051786417],
-                        [-73.63918895155356, 45.45831781859765],
-                        [-73.63925600678861, 45.45834274678629],
-                        [-73.6392459485059, 45.458355916390136],
-                        [-73.63928819329325, 45.458372378390624],
-                        [-73.63929154605415, 45.458369556333746],
-                        [-73.63930495709774, 45.45837378941903],
-                        [-73.63930093378467, 45.45837708181846],
-                        [-73.63935055464597, 45.45839542518324],
-                        [-73.63935390740689, 45.458392603127524],
-                        [-73.63936597734612, 45.45839683621106],
-                        [-73.6393632951374, 45.45840106929428],
-                        [-73.63940553992472, 45.45841517956941],
-                        [-73.63941090434217, 45.458409065117305],
-                        [-73.63942230372923, 45.458411416829715],
-                        [-73.63944107919026, 45.45838366661676],
-                        [-73.63958860069478, 45.45843822634611],
-                        [-73.6395946356644, 45.45845515866468],
-                        [-73.639574519099, 45.458481027474914],
-                        [-73.63959597676876, 45.45848902328656],
-                        [-73.63958994179914, 45.458497489438866],
-                        [-73.63966437309112, 45.45852429891275],
-                        [-73.63968516020869, 45.45856474827028],
-                        [-73.63962346940814, 45.45864282485506],
-                        [-73.63957049575275, 45.45865344331325]
-                    ]
-                ]
-            },
-            properties: {
-                name: "F. C. Smith Building",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.6396, 45.45861]
-            }
-        },
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                "coordinates": [
-                    [
-                        [-73.63805552075705, 45.45700768235533],
-                        [-73.6379256031108, 45.456957793649636],
-                        [-73.63787154978354, 45.45702630746083],
-                        [-73.63710103637098, 45.456725035496945],
-                        [-73.63736940639934, 45.45638978086584],
-                        [-73.63814132591384, 45.45669310657423],
-                        [-73.63806072007493, 45.45679487996479],
-                        [-73.6381868445052, 45.45684343844589],
-                        [-73.63805552075705, 45.45700768235533]
-                    ]
-                ]
-            },
-            properties: {
-                name: "Concordia Athletics Sports Camps (RA)",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.6379, 45.4569]
-            }
-        }, {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                "coordinates": [
-                    [
-                        [-73.63762943293287, 45.45728339527363],
-                        [-73.63677581997406, 45.45694944385062],
-                        [-73.63698905556116, 45.45668181092982],
-                        [-73.63784199798226, 45.45701576393934],
-                        [-73.63762943293287, 45.45728339527363]
-                    ]
-                ]
-            },
-            properties: {
-                name: "Concordia University School of Health (PC)",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.6377, 45.4571]
-            }
-        },
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.64057596908397, 45.45922624583542],
-                        [-73.64015553285907, 45.459060687137764],
-                        [-73.64025276292901, 45.45893745847219],
-                        [-73.64023264636361, 45.45893040338791],
-                        [-73.64041973043248, 45.458696644426546],
-                        [-73.64043180037172, 45.45870087749478],
-                        [-73.64045929301109, 45.45866748328123],
-                        [-73.6405357359596, 45.45869711476748],
-                        [-73.64055317031628, 45.458674068057306],
-                        [-73.64077378201259, 45.458759670077555],
-                        [-73.64076037096899, 45.45877613198937],
-                        [-73.64083882557405, 45.458806704098556],
-                        [-73.64059566166908, 45.459110285466956],
-                        [-73.6405983438778, 45.459126747276414],
-                        [-73.64064058866514, 45.459143679418254],
-                        [-73.64062382486064, 45.459164374251394],
-                        [-73.64063388314334, 45.45917660301283],
-                        [-73.64061108436923, 45.45920435288479],
-                        [-73.64059297946037, 45.45920482322147],
-                        [-73.64057596908397, 45.45922624583542]
-                    ]
-                ]
-            },
-            properties: {
-                name: "Concordia Applied Psychology Centre (PY)",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.64058, 45.45915]
-            }
-        },
-
-
-        {
-            type: "Feature",
-            geometry: {
-                type: "Polygon",
-                coordinates: [
-                    [
-                        [-73.64330853865712, 45.458554499011126],
-                        [-73.64323544846951, 45.45852486749199],
-                        [-73.64322069632155, 45.45854415149829],
-                        [-73.64311944294238, 45.45850323176976],
-                        [-73.64313084232946, 45.45849006219564],
-                        [-73.64305506993311, 45.4584590196159],
-                        [-73.64309999692917, 45.4584049302316],
-                        [-73.6431127374206, 45.45841057434368],
-                        [-73.64312212515111, 45.4583988157762],
-                        [-73.6430959736161, 45.45838846823477],
-                        [-73.64315364110357, 45.458317446421934],
-                        [-73.64322807239554, 45.458347078050086],
-                        [-73.64324215399132, 45.458329675349735],
-                        [-73.64334206626613, 45.458369654518226],
-                        [-73.64332865522252, 45.4583842351491],
-                        [-73.64340174541015, 45.45841480742739],
-                        [-73.64334810123574, 45.45847924432889],
-                        [-73.64336486504024, 45.458486299459605],
-                        [-73.64330853865712, 45.458554499011126]
-                    ]
-
-
-                ]
-            },
-            properties: {
-                name: "Jesuit Residence (JR)",
-                color: theme.colors.primaryLight,
-                textPosition: [-73.6432, 45.4586]
-            }
-        },
-    ]
-};
