+++ eslint.config.js
@@ -15,13 +15,14 @@ module.exports = [
             },
         },
         rules: {
-            "@typescript-eslint/no-unused-vars": ["off"],
+            "@typescript-eslint/no-unused-vars": ["warn", { argsIgnorePattern: "^_" }],
             "@typescript-eslint/no-explicit-any": "error",
             "@typescript-eslint/explicit-function-return-type": [
                 "warn",
                 { allowExpressions: true },
             ],
             "@typescript-eslint/no-empty-function": "warn",
+            semi: ["off"],
 
         },
         plugins: {
@@ -37,15 +38,15 @@ module.exports = [
             parserOptions: {
                 ecmaVersion: "latest",
                 sourceType: "module",
-                requireConfigFile: false,
+                requireConfigFile: false, // Ensures Babel works without an extra config file
                 babelOptions: {
-                    presets: ["@babel/preset-react"],
+                    presets: ["@babel/preset-react"], // Enables JSX parsing
                 },
             },
         },
         rules: {
-            "no-unused-vars": ["off"],
-            "no-console": ["warn", { allow: ["warn", "error"] }],
+            "no-unused-vars": ["warn", { argsIgnorePattern: "^_" }],
+            "no-console": "warn",
             "semi": ["off"],
         },
     },
