+++ package.json
@@ -12,8 +12,7 @@
     "ios": "expo run:ios",
     "web": "expo start --web",
     "test": "jest",
-    "lint": "eslint . --max-warnings=0",
-    "prettier:check": "prettier --check .",
+    "lint": "expo lint",
     "type-check": "tsc --noEmit"
   },
   "jest": {
@@ -24,19 +23,14 @@
     "@react-native-async-storage/async-storage": "1.23.1",
     "@react-navigation/bottom-tabs": "^7.2.0",
     "@react-navigation/native": "^7.0.14",
-    "@rnmapbox/maps": "^10.1.33",
     "expo": "~52.0.26",
     "expo-auth-session": "^6.0.3",
     "expo-blur": "^14.0.3",
-    "expo-camera": "^16.0.14",
     "expo-constants": "~17.0.4",
     "expo-font": "~13.0.3",
     "expo-haptics": "~14.0.1",
     "expo-linking": "~7.0.4",
-    "expo-media-library": "~17.0.5",
     "expo-router": "~4.0.17",
-    "expo-speech": "~13.0.1",
-    "expo-speech-recognition": "^1.1.1",
     "expo-splash-screen": "~0.29.21",
     "expo-status-bar": "~2.0.1",
     "expo-symbols": "~0.2.1",
@@ -47,12 +41,9 @@
     "react-native": "0.76.6",
     "react-native-calendars": "^1.1308.1",
     "react-native-gesture-handler": "~2.20.2",
-    "react-native-get-random-values": "^1.11.0",
-    "react-native-google-places-autocomplete": "^2.5.7",
     "react-native-reanimated": "~3.16.1",
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.4.0",
-    "react-native-vector-icons": "^10.2.0",
     "react-native-web": "~0.19.13",
     "react-native-webview": "13.12.5"
   },
