+++ services/userService.js
@@ -1,3 +1,4 @@
+import React from 'react';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
 export const getUserInfo = async (token) => {
@@ -12,7 +13,7 @@ export const getUserInfo = async (token) => {
     await AsyncStorage.setItem("@user", JSON.stringify(user));
     return user;
     }catch(error){
-      console.warn(error);
+      console.log(error);
       return null;
     }
   }
